autosave_guardian/requirements.txt
Flask==3.0.3
PyJWT==2.9.0
gunicorn==22.0.0
Flask-Cors==4.0.1
PyJWT==2.9.0
pymongo==4.10.1
pydantic==2.8.2
python-dotenv==1.0.1









